+++ jaxlib/gpu_solver.py
@@ -369,7 +369,14 @@ def _gesvd_hlo(platform, gpu_solver, have_jacobi_solver, dtype, a,
   vector_layout = (num_bd,) + tuple(range(num_bd - 1, -1, -1))
   i32_type = ir.IntegerType.get_signless(32)
 
-  if have_jacobi_solver and m < 32 and n < 32:
+  # NVIDIA's batched Jacobi solver supports a maximum matrix size of 32x32, but
+  # the unbatched solver has no such limit. The unbatched solver appears to
+  # outperform gesvd for small-moderate matrices, e.g., see:
+  # https://developer.download.nvidia.com/video/gputechconf/gtc/2019/presentation/s9226-fast-singular-value-decomposition-on-gpus-v2.pdf
+  # slide 5.
+  if have_jacobi_solver and (
+      (b == 1 and m <= 1024 and n <= 1024) or (m <= 32 and n <= 32)
+  ):
     # The batched kernel doesn't support "econ" mode.
     econ = not full_matrices and b == 1
     lwork, opaque = gpu_solver.build_gesvdj_descriptor(